various improvements based on PR feedback
