+++ src/shared/cloudformation/activation.ts
@@ -5,14 +5,13 @@
 
 import * as vscode from 'vscode'
 import { getLogger } from '../logger'
-import { activateExtension, localize } from '../utilities/vsCodeUtils'
+import { localize } from '../utilities/vsCodeUtils'
 
 import { CloudFormationTemplateRegistry } from './templateRegistry'
 import { ext } from '../extensionGlobals'
 import { getIdeProperties } from '../extensionUtilities'
 import { NoopWatcher } from '../watchedFiles'
 import { createStarterTemplateFile } from './cloudformation'
-import { VSCODE_EXTENSION_ID } from '../extensions'
 
 export const TEMPLATE_FILE_GLOB_PATTERN = '**/*.{yaml,yml}'
 
@@ -31,9 +30,6 @@ export const TEMPLATE_FILE_EXCLUDE_PATTERN = /.*[/\\]\.aws-sam([/\\].*|$)/
  * @param extensionContext VS Code extension context
  */
 export async function activate(extensionContext: vscode.ExtensionContext): Promise<void> {
-    // Note: redhat.vscode-yaml no longer works on vscode 1.42
-    activateExtension(VSCODE_EXTENSION_ID.yaml).then(addCustomTags)
-
     try {
         const registry = new CloudFormationTemplateRegistry()
         await registry.addExcludedPattern(TEMPLATE_FILE_EXCLUDE_PATTERN)
@@ -58,57 +54,4 @@ export async function activate(extensionContext: vscode.ExtensionContext): Promi
         vscode.commands.registerCommand('aws.cloudFormation.newTemplate', () => createStarterTemplateFile(false)),
         vscode.commands.registerCommand('aws.sam.newTemplate', () => createStarterTemplateFile(true))
     )
-}
-
-/**
- * Adds custom tags to the YAML extension's settings in order to hide error
- * notifications for SAM/CFN intrinsic functions if a user has the YAML extension.
- *
- * Lifted near-verbatim from the cfn-lint VSCode extension.
- * https://github.com/aws-cloudformation/cfn-lint-visual-studio-code/blob/629de0bac4f36cfc6534e409a6f6766a2240992f/client/src/extension.ts#L56
- */
-function addCustomTags(): void {
-    const settingName = 'yaml.customTags'
-    const currentTags = vscode.workspace.getConfiguration().get<string[]>(settingName) ?? []
-    if (!Array.isArray(currentTags)) {
-        getLogger().error(
-            'setting "%s" is not an array. SAM/CFN intrinsic functions will not be recognized.',
-            settingName
-        )
-        return
-    }
-    const cloudFormationTags = [
-        '!And',
-        '!And sequence',
-        '!If',
-        '!If sequence',
-        '!Not',
-        '!Not sequence',
-        '!Equals',
-        '!Equals sequence',
-        '!Or',
-        '!Or sequence',
-        '!FindInMap',
-        '!FindInMap sequence',
-        '!Base64',
-        '!Join',
-        '!Join sequence',
-        '!Cidr',
-        '!Ref',
-        '!Sub',
-        '!Sub sequence',
-        '!GetAtt',
-        '!GetAZs',
-        '!ImportValue',
-        '!ImportValue sequence',
-        '!Select',
-        '!Select sequence',
-        '!Split',
-        '!Split sequence',
-    ]
-    const missingTags = cloudFormationTags.filter(item => !currentTags.includes(item))
-    if (missingTags.length > 0) {
-        const updateTags = currentTags.concat(missingTags)
-        vscode.workspace.getConfiguration().update('yaml.customTags', updateTags, vscode.ConfigurationTarget.Global)
-    }
 }
